Splited out request data by their suite
