+++ integration-tests/constants/requestData/endToEndFlowRequestData.ts
@@ -1,42 +1,23 @@
-import { ZendeskRequestData } from '../types/zendeskRequestData'
+import { ZendeskRequestData } from '../../types/zendeskRequestData'
+import { generateRandomNumber } from '../../utils/helpers'
 import {
-  generateRandomNumber,
-  generateZendeskRequestDate
-} from '../utils/helpers'
+  END_TO_END_TEST_DATA_PATH,
+  END_TO_END_TEST_DATE
+} from '../awsParameters'
 import {
-  INTEGRATION_TEST_DATE,
-  TEST_DATA_DATA_PATHS,
-  INTEGRATION_TEST_DATE_GLACIER,
-  TEST_DATA_EVENT_ID,
-  INTEGRATION_TEST_DATE_MIX_DATA,
-  INTEGRATION_TEST_DATE_NO_DATA
-} from './awsParameters'
-import {
+  END_TO_END_TEST_EVENT_ID,
+  END_TO_END_TEST_JOURNEY_ID,
+  END_TO_END_TEST_SESSION_ID,
+  END_TO_END_TEST_USER_ID,
   ZendeskFormFieldIDs,
   ZENDESK_END_USER_EMAIL,
   ZENDESK_END_USER_NAME,
   ZENDESK_PII_FORM_ID
-} from './zendeskParameters'
+} from '../zendeskParameters'
 
-const generateSubjectLine = () => {
-  const fixedSubjectLine = process.env.FIXED_SUBJECT_LINE
-  if (fixedSubjectLine) {
-    return fixedSubjectLine
-  }
-  return `Integration Test Request - ${generateRandomNumber()}`
-}
-
-const generateEventIds = () => {
-  const overrideEventIds = process.env.OVERRIDE_EVENT_IDS
-  if (overrideEventIds) {
-    return overrideEventIds
-  }
-  return TEST_DATA_EVENT_ID
-}
-
-export const validRequestData: ZendeskRequestData = {
+export const endToEndFlowRequestDataWithEventId: ZendeskRequestData = {
   request: {
-    subject: generateSubjectLine(),
+    subject: `Integration Test Request - ` + generateRandomNumber(),
     ticket_form_id: ZENDESK_PII_FORM_ID,
     custom_fields: [
       {
@@ -45,27 +26,23 @@ export const validRequestData: ZendeskRequestData = {
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_EVENT_ID_LIST_FIELD_ID,
-        value: generateEventIds()
+        value: END_TO_END_TEST_EVENT_ID
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
-        value: INTEGRATION_TEST_DATE
+        value: END_TO_END_TEST_DATE
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
-        value: ['drivers_license']
+        value: null
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_CUSTOM_DATA_PATH_FIELD_ID,
-        value: process.env.DATA_PATHS
-          ? process.env.DATA_PATHS
-          : TEST_DATA_DATA_PATHS
+        value: END_TO_END_TEST_DATA_PATH
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_EMAIL,
-        value: process.env.FIXED_RECIPIENT_EMAIL
-          ? process.env.FIXED_RECIPIENT_EMAIL
-          : ZENDESK_END_USER_EMAIL
+        value: ZENDESK_END_USER_EMAIL
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_NAME,
@@ -78,35 +55,102 @@ export const validRequestData: ZendeskRequestData = {
   }
 }
 
-export const setCustomFieldValueForRequest = (
-  requestData: ZendeskRequestData,
-  fieldId: number,
-  fieldValue: string
-) => {
-  const customField = requestData.request.custom_fields.find(
-    (f) => f.id === fieldId
-  )
-  if (customField) {
-    customField.value = fieldValue
-  }
-}
-
-export const invalidRequestData: ZendeskRequestData = {
+export const endToEndFlowRequestDataWithUserId: ZendeskRequestData = {
   request: {
     subject: `Integration Test Request - ` + generateRandomNumber(),
     ticket_form_id: ZENDESK_PII_FORM_ID,
     custom_fields: [
       {
         id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_FIELD_ID,
-        value: 'event_id'
+        value: 'user_id'
       },
       {
-        id: ZendeskFormFieldIDs.PII_FORM_EVENT_ID_LIST_FIELD_ID,
-        value: '637783 3256'
+        id: ZendeskFormFieldIDs.PII_FORM_USER_ID_LIST_FIELD_ID,
+        value: END_TO_END_TEST_USER_ID
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
-        value: generateZendeskRequestDate(50)
+        value: END_TO_END_TEST_DATE
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
+        value: ['passport_number', 'passport_expiry_date']
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_CUSTOM_DATA_PATH_FIELD_ID,
+        value: ''
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_EMAIL,
+        value: ZENDESK_END_USER_EMAIL
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_NAME,
+        value: ZENDESK_END_USER_NAME
+      }
+    ],
+    comment: {
+      body: 'PII request created in integration test'
+    }
+  }
+}
+
+export const endToEndFlowRequestDataWithSessionId: ZendeskRequestData = {
+  request: {
+    subject: `Integration Test Request - ` + generateRandomNumber(),
+    ticket_form_id: ZENDESK_PII_FORM_ID,
+    custom_fields: [
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_FIELD_ID,
+        value: 'session_id'
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_SESSION_ID_LIST_FIELD_ID,
+        value: END_TO_END_TEST_SESSION_ID
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
+        value: END_TO_END_TEST_DATE
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
+        value: ['name', 'dob', 'addresses']
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_CUSTOM_DATA_PATH_FIELD_ID,
+        value: ''
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_EMAIL,
+        value: ZENDESK_END_USER_EMAIL
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_NAME,
+        value: ZENDESK_END_USER_NAME
+      }
+    ],
+    comment: {
+      body: 'PII request created in integration test'
+    }
+  }
+}
+
+export const endToEndFlowRequestDataWithJourneyId: ZendeskRequestData = {
+  request: {
+    subject: `Integration Test Request - ` + generateRandomNumber(),
+    ticket_form_id: ZENDESK_PII_FORM_ID,
+    custom_fields: [
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_FIELD_ID,
+        value: 'journey_id'
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_JOURNEY_ID_LIST_FIELD_ID,
+        value: END_TO_END_TEST_JOURNEY_ID
+      },
+      {
+        id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
+        value: END_TO_END_TEST_DATE
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
@@ -131,7 +175,7 @@ export const invalidRequestData: ZendeskRequestData = {
   }
 }
 
-export const validApiTestRequestData = {
+export const endToEndFlowRequestDataNoMatch: ZendeskRequestData = {
   request: {
     subject: `Integration Test Request - ` + generateRandomNumber(),
     ticket_form_id: ZENDESK_PII_FORM_ID,
@@ -142,19 +186,19 @@ export const validApiTestRequestData = {
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_EVENT_ID_LIST_FIELD_ID,
-        value: TEST_DATA_EVENT_ID
+        value: 'zzzzzzzz-yyyy-aaaa-bbbb-cccccccccccc'
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
-        value: generateZendeskRequestDate(-60)
+        value: END_TO_END_TEST_DATE
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
-        value: ['drivers_license']
+        value: null
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_CUSTOM_DATA_PATH_FIELD_ID,
-        value: TEST_DATA_DATA_PATHS
+        value: END_TO_END_TEST_DATA_PATH
       },
       {
         id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_EMAIL,
@@ -167,153 +211,6 @@ export const validApiTestRequestData = {
     ],
     comment: {
       body: 'PII request created in integration test'
-    }
-  }
-}
-
-export const validGlacierRequestData: ZendeskRequestData = {
-  request: {
-    subject: `Integration Test Request - ` + generateRandomNumber(),
-    ticket_form_id: ZENDESK_PII_FORM_ID,
-    custom_fields: [
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_FIELD_ID,
-        value: 'event_id'
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_EVENT_ID_LIST_FIELD_ID,
-        value: TEST_DATA_EVENT_ID
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
-        value: INTEGRATION_TEST_DATE_GLACIER
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
-        value: ['drivers_license']
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_CUSTOM_DATA_PATH_FIELD_ID,
-        value: ''
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_EMAIL,
-        value: ZENDESK_END_USER_EMAIL
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_NAME,
-        value: ZENDESK_END_USER_NAME
-      }
-    ],
-    comment: {
-      body: 'PII request created in integration test'
-    }
-  }
-}
-
-export const validStandardAndGlacierTiersRequestData: ZendeskRequestData = {
-  request: {
-    subject: `Integration Test Request - ` + generateRandomNumber(),
-    ticket_form_id: ZENDESK_PII_FORM_ID,
-    custom_fields: [
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_FIELD_ID,
-        value: 'event_id'
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_EVENT_ID_LIST_FIELD_ID,
-        value: TEST_DATA_EVENT_ID
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
-        value: INTEGRATION_TEST_DATE_MIX_DATA
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
-        value: ['drivers_license']
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_CUSTOM_DATA_PATH_FIELD_ID,
-        value: ''
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_EMAIL,
-        value: ZENDESK_END_USER_EMAIL
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_NAME,
-        value: ZENDESK_END_USER_NAME
-      }
-    ],
-    comment: {
-      body: 'PII request created in integration test'
-    }
-  }
-}
-
-export const validRequestNoData: ZendeskRequestData = {
-  request: {
-    subject: `Integration Test Request - ` + generateRandomNumber(),
-    ticket_form_id: ZENDESK_PII_FORM_ID,
-    custom_fields: [
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_FIELD_ID,
-        value: 'event_id'
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_EVENT_ID_LIST_FIELD_ID,
-        value: TEST_DATA_EVENT_ID
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUEST_DATE_FIELD_ID,
-        value: INTEGRATION_TEST_DATE_NO_DATA
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUESTED_PII_TYPE_FIELD_ID,
-        value: ['drivers_license']
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_CUSTOM_DATA_PATH_FIELD_ID,
-        value: ''
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_EMAIL,
-        value: ZENDESK_END_USER_EMAIL
-      },
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_IDENTIFIER_RECIPIENT_NAME,
-        value: ZENDESK_END_USER_NAME
-      }
-    ],
-    comment: {
-      body: 'PII request created in integration test'
-    }
-  }
-}
-
-export const ticketApprovalData = {
-  ticket: {
-    tags: ['process_started', 'approved'],
-    custom_fields: [
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUEST_STATUS_FIELD_ID,
-        value: 'approved'
-      }
-    ],
-    status: 'open',
-    fields: [
-      {
-        id: ZendeskFormFieldIDs.PII_FORM_REQUEST_STATUS_FIELD_ID,
-        value: 'approved'
-      }
-    ],
-    collaborator_ids: [],
-    follower_ids: [],
-    comment: {
-      body: '<p>Request <b>APPROVED</b> and data retrieval has started...</p>',
-      html_body:
-        '<p>Request <b>APPROVED</b> and data retrieval has started...</p>',
-      public: 'true'
     }
   }
 }
