+++ integration-tests/utils/zendesk/approveZendeskTicket.ts
@@ -1,6 +1,6 @@
 import axios, { AxiosPromise } from 'axios'
 import { authoriseAs } from './authoriseAs'
-import { ticketApprovalData } from '../../constants/requestData'
+import { ticketApprovalData } from '../../constants/requestData/dataCopyRequestData'
 import {
   ZENDESK_AGENT_EMAIL,
   ZENDESK_BASE_URL,
