+++ integration-tests/zendeskWebhookAPI.spec.ts
@@ -2,7 +2,6 @@ import axios from 'axios'
 import { deleteZendeskTicket } from './utils/zendesk/deleteZendeskTicket'
 import { generateZendeskRequestDate, getEnv } from './utils/helpers'
 import { createZendeskTicket } from './utils/zendesk/createZendeskTicket'
-import { validApiTestRequestData } from './constants/requestData'
 import {
   ZENDESK_END_USER_EMAIL,
   ZENDESK_END_USER_NAME
@@ -13,6 +12,7 @@ import {
   TEST_DATA_DATA_PATHS,
   TEST_DATA_EVENT_ID
 } from './constants/awsParameters'
+import { validApiTestRequestData } from './constants/requestData/webhookAPIRequestData'
 
 const webhookUrl = `${getEnv('ZENDESK_WEBHOOK_API_BASE_URL')}/zendesk-webhook`
